+++ CHANGELOG.md
@@ -14,11 +14,13 @@ Keep it human-readable, your future self will thank you!
 
 - Fix: Enable inference when no constant forcings are used
 - Add anemoi-transform link to documentation
+- Add support for unstructured grids
 
 ### Changed
 
 - Add cos_solar_zenith_angle to list of known forcings
 - Add missing classes in checkpoint handling
+- Rename Condition to State [#24](https://github.com/ecmwf/anemoi-inference/pull/24)
 - Fix pre-commit regex
 
 ### Removed